Revert "Merge branch 'iNavFlight:master' into release_5.1.0"
This reverts commit dddab3b5d8274d7ecabb309566f2a57eb90c4e36, reversing
changes made to c62200fb2481c7b91acd8c3c2d359ebde30a8d5b.

+++ src/main/io/osd_common.h
@@ -45,7 +45,7 @@ typedef struct {
 
 PG_DECLARE(osdCommonConfig_t, osdCommonConfig);
 
-int16_t osdGetSpeedFromSelectedSource(void);
+int osdGetSpeedFromSelectedSource(void);
 
 #endif // defined(USE_OSD) || defined(USE_DJI_HD_OSD)
